agents/adk/agents/software-bug-assistant/pyproject.toml
[project]
name = "software-bug-assistant"
version = "0.1.0"
description = "Software bug assistant using ADK"
authors = [{ name = "Jack Wotherspoon", email = "jackwoth@google.com" }]
license = {text = "Apache License 2.0"}
readme = "README.md"
requires-python = ">=3.10,<3.14"

dependencies = [
    "google-adk>=1.8.0",
    "langchain>=0.3.0,<1.3.0",
    "google-cloud-aiplatform[agent-engines,evaluation]>=1.93.0",
    "langchain-community>=0.3.25",
    "python-dotenv>=1.1.0",
    "stackapi>=0.3.1",
    "toolbox-core>=0.1.0",
    "a2a-sdk>=0.3.0",
    "starlette>=0.46.1",
    "uvicorn>=0.34.0",
    "click>=8.1.8",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "agent-starter-pack>=0.14.1",
    "ruff>=0.4.6",
    "mypy>=1.15.0",
    "codespell>=2.2.0",
    "types-pyyaml>=6.0.12.20240917",
    "types-requests>=2.32.0.20240914",
]

[tool.ruff]
# Point back to the Master template in the root
extend = "../../../pyproject.toml"

[tool.ruff.lint.isort]
# Categorizes the local package for correct import sorting
known-first-party = ["software_bug_assistant"]

[tool.mypy]
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
check_untyped_defs = true
disallow_subclassing_any = true
warn_incomplete_stub = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
follow_imports = "silent"
ignore_missing_imports = true
explicit_package_bases = true
disable_error_code = ["misc", "no-untyped-call", "no-any-return"]
exclude = [".venv"]

[tool.hatch.build.targets.wheel]
# Ensures Hatchling finds the package folder correctly
packages = ["software_bug_assistant"]

[tool.pytest.ini_options]
pythonpath = "."
asyncio_default_fixture_loop_scope = "function"

[tool.agent-starter-pack]
example_question = "Can you help me debug a login page issue?"

[tool.agent-starter-pack.settings]
agent_directory = "software_bug_assistant"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"